pyproject.toml
[project]
name = "mock-oauth2-service"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Michael Dryden", email = "mjdryden@gmail.com" }]
dependencies = [
    "FastAPI>=0.109.2",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "uvicorn>=0.27.0.post1",
    "python-multipart>=0.0.7",
    "cryptography>=42.0.2",
    "strivelogger>=1.1.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ["ruff>=0.2.1"]

[tool.rye.scripts]
dev = "uvicorn app.main:app --reload --host 0.0.0.0"
lint = "ruff check . --output-format=github"
format = "ruff format . --check"


[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 160

[tool.ruff]
line-length = 168

[tool.pyright]
include = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
